refactor: 기존 도메인 도메인패키지로 이동
