+++ src/main/java/roomescape/domain/theme/ThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.theme;
+package roomescape.domain.theme;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.entities.Theme;
-import roomescape.theme.data.ThemeAddRequestDto;
+import roomescape.domain.theme.data.ThemeAddRequestDto;
 
 import java.util.List;
